Rename profile button and move to shared components
