+++ src/Patient/ContactTracing/index.js
@@ -5,7 +5,8 @@ import Typography from '@material-ui/core/Typography';
 import OverTopBar from '../Navigation/OverTopBar';
 import XIcon from '@material-ui/icons/Clear';
 import { useTranslation } from 'react-i18next';
-import ProfileButton from '../../Practitioner/PatientProfile/ProfileButton';
+import ProfileButton from '../../Components/FlatButton';
+import Grid from '@material-ui/core/Grid'
 
 const useStyles = makeStyles({
     body: {
@@ -23,7 +24,9 @@ const ContactTracingUpdate = () => {
         <OverTopBar notFixed handleBack={() => { uiStore.push("/") }} title={t('updatedContactTracing.title')} />
         <div className={classes.body}>
             <p>Hello</p>
-            <ProfileButton>Submit</ProfileButton>
+            <Grid justify="flex-end" container spacing={1}>
+                <ProfileButton>{t('coordinator.patientProfile.editDetails.submit')}</ProfileButton>
+            </Grid>
         </div>
     </div>)
