+++ src/Patient/Home/PushEnrollmentReminder.js
@@ -4,7 +4,7 @@ import Typography from '@material-ui/core/Typography'
 import Grid from '@material-ui/core/Grid'
 import Colors from '../../Basics/Colors'
 import WarningIcon from '@material-ui/icons/WarningRounded'
-import ProfileButton from '../../Practitioner/PatientProfile/ProfileButton'
+import ProfileButton from '../../Components/FlatButton'
 import RightIcon from '@material-ui/icons/KeyboardArrowRight'
 import useStores from '../../Basics/UseStores'
 import { useTranslation } from 'react-i18next'
